Display network as JavaScript function
The visualization is great! But I think it's also unclear what exactly it's computing.
E.g. where do the weights go in the calculation?
Does the bias go before or after the activation function?
For my own understanding, I found it helpful to show the network as a JavaScript function.
This makes the semantics of the network clear.